@@ -1,5 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from 'react'
+import PropTypes from 'prop-types'
 import {
   withStyles,
   ListItem,
@@ -9,8 +9,9 @@ import {
   Avatar,
   IconButton,
   Switch,
-} from '@material-ui/core';
-import { Folder, CalendarToday, Timer } from '@material-ui/icons';
+} from '@material-ui/core'
+import { Folder, CalendarToday, Timer as TimerIcon } from '@material-ui/icons'
+import Timer from './Timer'
 
 const styles = theme => ({
   item: {
@@ -18,63 +19,51 @@ const styles = theme => ({
     minHeight: 65,
     margin: `${theme.spacing.unit}px 0 0`,
   },
-});
+})
 
-const fixGpio = gpio => (gpio < 10 ? `0${gpio}` : gpio);
+const fixGpio = gpio => (gpio < 10 ? `0${gpio}` : gpio)
 
-class InteractiveList extends React.Component {
-  state = {
-    switch: false,
-  };
+const Sensor = props => {
+  const [enable, setEnable] = useState(false)
 
-  handleSwitch = gpio => {
-    this.setState({ switch: !this.state.switch });
-
+  const handleSwitch = gpio => {
     const url = `http://192.168.0.155:5000/api/turnOnRelay?gpio=${fixGpio(
       gpio,
-    )}`;
-    console.log('url', url);
+    )}`
 
     fetch(url)
       .then(response => {
-        console.log('YAYYY', response);
+        setEnable(!enable)
+        console.log('response', response)
       })
-      .catch(e => console.log(e));
-  };
-
-  componentDidMount = () => {};
-
-  render() {
-    const { classes, item } = this.props;
-
-    return (
-      <ListItem button className={classes.item}>
-        <ListItemAvatar>
-          <Avatar>
-            <Folder />
-          </Avatar>
-        </ListItemAvatar>
-        <ListItemText primary={item.name} secondary={item.desc} />
-        <ListItemSecondaryAction>
-          <IconButton className={classes.button} aria-label="Calendar">
-            <CalendarToday />
-          </IconButton>
-          <IconButton className={classes.button} aria-label="Timer">
-            <Timer />
-          </IconButton>
-          <Switch
-            onChange={this.handleSwitch.bind(this, item.gpio)}
-            checked={this.state.switch}
-          />
-        </ListItemSecondaryAction>
-      </ListItem>
-    );
+      .catch(e => console.log(e))
   }
+  const { classes, item } = props
+  return (
+    <ListItem button className={classes.item}>
+      <ListItemAvatar>
+        <Avatar>
+          <Folder />
+        </Avatar>
+      </ListItemAvatar>
+      <ListItemText primary={item.name} secondary={item.desc} />
+      <ListItemSecondaryAction>
+        <IconButton className={classes.button} aria-label="Calendar">
+          <CalendarToday />
+        </IconButton>
+        <IconButton className={classes.button} aria-label="Timer">
+          <TimerIcon />
+        </IconButton>
+        <Timer callBack={() => console.log('Timerrr')} {...props} />
+        <Switch onChange={() => handleSwitch(item.gpio)} checked={enable} />
+      </ListItemSecondaryAction>
+    </ListItem>
+  )
 }
 
-InteractiveList.propTypes = {
+Sensor.propTypes = {
   classes: PropTypes.object.isRequired,
   item: PropTypes.object.isRequired,
-};
+}
 
-export default withStyles(styles)(InteractiveList);
+export default withStyles(styles)(Sensor)
