@@ -1,20 +1,6 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import {
-  withStyles,
-  List,
-  ListItem,
-  ListItemAvatar,
-  ListItemSecondaryAction,
-  ListItemText,
-  Avatar,
-  IconButton,
-  Grid,
-  Typography,
-  Switch,
-  Icon,
-} from '@material-ui/core';
-import { Folder, CalendarToday, Timer } from '@material-ui/icons';
+import React from 'react'
+import PropTypes from 'prop-types'
+import { withStyles, List, Grid, Typography } from '@material-ui/core'
 
 const styles = theme => ({
   root: {
@@ -30,30 +16,28 @@ const styles = theme => ({
     minHeight: 65,
     margin: `${theme.spacing.unit}px 0 0`,
   },
-});
+})
 
-class InteractiveList extends React.Component {
-  render() {
-    const { classes, title, children } = this.props;
-
-    return (
-      <div className={classes.root}>
-        <Grid container spacing={0} justify="center">
-          <Grid md={8}>
-            <Typography variant="h6" className={classes.title}>
-              {title}
-            </Typography>
-            <List>{children}</List>
-          </Grid>
+const InteractiveList = props => {
+  const { classes, title, children } = props
+  return (
+    <div className={classes.root}>
+      <Grid container spacing={0} justify="center">
+        <Grid item md={8}>
+          <Typography variant="h6" className={classes.title}>
+            {title}
+          </Typography>
+          <List>{children}</List>
         </Grid>
-      </div>
-    );
-  }
+      </Grid>
+    </div>
+  )
 }
 
 InteractiveList.propTypes = {
   classes: PropTypes.object.isRequired,
-  Title: PropTypes.string.isRequired,
-};
+  title: PropTypes.string.isRequired,
+  children: PropTypes.array.isRequired,
+}
 
-export default withStyles(styles)(InteractiveList);
+export default withStyles(styles)(InteractiveList)
