@@ -1,2 +1,2 @@
-export const LIST_SENSOR_TITLE = 'Sensores';
-export const PAGE_TITLE = 'General';
+export const LIST_SENSOR_TITLE = 'Sensores'
+export const PAGE_TITLE = 'General'
